refactored helpers code
added ignore in teardwon for apps not created in the reservation
